CRUD for posts done

+++ backend/api/posts/postsController.js
@@ -27,17 +27,50 @@ export const addOnePostController = asyncHandler(async (req, res) => {
 })
 
 export const updateOnePostController = asyncHandler(async (req, res) => {
-    //JWT validation
-    //Has to allow change to single or multiple parameters
-    const response = await updateOnePostView()
-    res.status(200).json({
-        data: { message: "Data updated successfully" }
+    const { id } = req.params
+    const row = (await getOnePostView(id))[0]
+
+    if (!row) {
+        const err = new Error("Post not found")
+        err.status = 404
+        throw err
+    }
+
+    if (req.user.userId != row.author_id) {
+        const err = new Error("You are not the author")
+        err.status = 403
+        throw err
+    }
+
+    // 2. merge 
+    const updatedPost = {
+        title: req.body.title ?? row.title,
+        description: req.body.description ?? row.description
+    }
+
+    // 3. update
+    const response = await updateOnePostView(id, updatedPost)
+
+    return res.status(200).json({
+        message: "Post updated",
+        data: updatedPost
     })
 })
 
 export const deleteOnePostController = asyncHandler(async (req, res) => {
-    //JWT validation
     let id = req.params.id
+    const row = (await getOnePostView(id))[0]
+
+    if (!row) {
+        const err = new Error("Post not found")
+        err.status = 404
+        throw err
+    }
+    if (req.user.userId != row.author_id) {
+        const err = new Error("You are not the author")
+        err.status = 403
+        throw err
+    }
     const response = await deleteOnePostView(id)
     res.status(200).json({
         data: { message: "Data deleted successfully" }
